+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.2.1
 Upgrade Checkout SDK to v4.
-Add `generateApplePayToken` method.
+Add `generateApplePayToken` and `generateGooglePayToken` methods.
 Confirm it works with Flutter 3.7.x
 
 ## 1.2.0